correcciones pendientes

+++ src/main/java/com/clinicabienestar/api/dto/DetalleFacturaDTO.java
@@ -1,10 +1,11 @@
 package com.clinicabienestar.api.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class DetalleFacturaDTO {
     private String descripcionServicio;
     private int cantidad;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
 }
